refactor: compile jsonschemas into lazy static refs
