+++ identity-wallet/src/jsonschemas/EDC_VerifiableCredentialV1_1.json
