+++ identity-wallet/src/jsonschemas/EuropeanDigitalCredentialV3_3.json
