+++ identity-wallet/src/jsonschemas/VerifiableCredentialV1_1.json
