+++ identity-wallet/src/jsonschemas/VerifiableCredentialV2.json
