+++ identity-wallet/src/state/did/validate_linked_verifiable_presentations.rs
@@ -1,5 +1,6 @@
+use crate::jsonschemas::validate_credential_types;
 use crate::state::{
-    core_utils::helpers::{download_logo, get_issuer_document, validate_credential_types},
+    core_utils::helpers::{download_logo, get_issuer_document},
     did::validate_domain_linkage::{ValidationStatus, Verifier},
 };
 use did_manager::Resolver;
@@ -64,7 +65,6 @@ pub async fn validate_linked_verifiable_presentations(
 
     info!("Holder document: {holder_document:#?}");
 
-    println!("2");
     iter(
         // Get all linked verifiable presentation URLs from the holder document
         holder_document
@@ -218,11 +218,9 @@ async fn get_validated_linked_credential_data(
                 ) {
                     info!("Validated linked verifiable credential JWT: {linked_verifiable_credential:#?}");
 
-                    println!("3");
                     // Validate the linked verifiable credential against its corresponding JSON Schema
                     validate_credential_types(&linked_verifiable_credential.credential.to_json_value().ok()?).ok()?;
 
-                    println!("4");
                     let credential_subject = match &linked_verifiable_credential.credential.credential_subject {
                         OneOrMany::One(subject) => Some(subject),
                         // TODO: how to handle multiple credential subjects?
@@ -757,7 +755,6 @@ mod tests {
 
         let resolver = Resolver::new().await;
 
-        println!("1");
         assert_eq!(
             validate_linked_verifiable_presentations(&resolver, holder.did_document.id().to_string().as_ref()).await,
             vec![
